Extract this out so it can be reused by `init` command.
